Moved remove transforms method to new controller.
