+++ conf/trusts.routes
@@ -51,7 +51,8 @@ PUT         /:identifier/trust-details/residential-status                       
 
 GET         /:identifier/taxable-migration/migrating-to-taxable                             controllers.TaxableMigrationController.getTaxableMigrationFlag(identifier: String)
 POST        /:identifier/taxable-migration/migrating-to-taxable                             controllers.TaxableMigrationController.setTaxableMigrationFlag(identifier: String)
-DELETE      /:identifier/transforms                                                         controllers.TaxableMigrationController.removeTransforms(identifier: String)
+
+DELETE      /:identifier/transforms                                                         controllers.transformations.TransformationController.removeTransforms(identifier: String)
 
 GET         /:identifier/years-returns                                                      controllers.GetTrustController.getYearsReturns(identifier: String)
 PUT         /:identifier/years-returns                                                      controllers.transformations.taxliability.TaxLiabilityTransformationController.setYearsReturns(identifier: String)
